requirements.txt
# Core dependencies
torch>=2.4.0
torchvision>=0.15.0
numpy>=1.23.0
opencv-python>=4.7.0
scipy>=1.10.0
tqdm>=4.65.0
pyyaml>=6.0

# Optional dependencies for development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

# Visualization and evaluation
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0

# Machine learning and evaluation
scikit-learn>=1.3.0
pandas>=1.5.0 